requirements.txt
google-cloud-bigquery
web3
